+++ include/rocksdb/options.h
@@ -1340,14 +1340,14 @@ struct DBOptions {
   // It indicates, which lowest cache tier we want to
   // use for a certain DB. Currently we support volatile_tier and
   // non_volatile_tier. They are layered. By setting it to kVolatileTier, only
-  // the block cache (current implemented volatile_tier) is used. So the block
+  // the block cache (current implemented volatile_tier) is used. So
   // cache entries will not spill to secondary cache (current
   // implemented non_volatile_tier), and block cache lookup misses will not
-  // lookup in the secondary cache. When it is set to kNonVolatileTier, we use
+  // lookup in the secondary cache. When kNonVolatileBlockTier is used, we use
   // both block cache and secondary cache.
   //
-  // Default: kNonVolatileTier
-  CacheTier lowest_used_cache_tier = CacheTier::kNonVolatileTier;
+  // Default: kNonVolatileBlockTier
+  CacheTier lowest_used_cache_tier = CacheTier::kNonVolatileBlockTier;
 };
 
 // Options to control the behavior of a database (passed to DB::Open)