+++ options/db_options.cc
@@ -43,7 +43,7 @@ static std::unordered_map<std::string, DBOptions::AccessHint>
 
 static std::unordered_map<std::string, CacheTier> cache_tier_string_map = {
     {"kVolatileTier", CacheTier::kVolatileTier},
-    {"kNonVolatileTier", CacheTier::kNonVolatileTier}};
+    {"kNonVolatileBlockTier", CacheTier::kNonVolatileBlockTier}};
 
 static std::unordered_map<std::string, InfoLogLevel> info_log_level_string_map =
     {{"DEBUG_LEVEL", InfoLogLevel::DEBUG_LEVEL},